windows fix for 2025 day 2

+++ 2025/day-02/main.cpp
@@ -9,8 +9,12 @@ using namespace std;
 // #define FILE "input/demo1.txt"
 #define FILE "input/input.txt"
 
-long part1() {
-  long n = 0;
+// on linux, 'long' worked fine
+// on windows 'long long' is needed
+// this is for numbers over 2^32
+// but only if unsigned (2^31 for signed type)
+long long part1() {
+  long long n = 0;
 
   string file;
   ifstream read_file(FILE);
@@ -19,11 +23,11 @@ long part1() {
     string segment;
 
     getline(range, segment, '-');
-    long start_i = stol(segment);
+    long long start_i = stoll(segment);
     getline(range, segment, '-');
-    long end_i = stol(segment);
+    long long end_i = stoll(segment);
 
-    for (long i = start_i; i <= end_i; i++) {
+    for (long long i = start_i; i <= end_i; i++) {
       string num = to_string(i);
       int len = num.length();
       if (len % 2 == 1) continue;
@@ -38,8 +42,8 @@ long part1() {
   return n;
 }
 
-long part2() {
-  long n = 0;
+long long part2() {
+  long long n = 0;
 
   string file;
   ifstream read_file(FILE);
@@ -48,31 +52,30 @@ long part2() {
     string segment;
 
     getline(range, segment, '-');
-    long start_i = stol(segment);
+    long long start_i = stoll(segment);
     getline(range, segment, '-');
-    long end_i = stol(segment);
+    long long end_i = stoll(segment);
 
-    for (long i = start_i; i <= end_i; i++) {
-      // cout << "A" << '\t' << i << endl;
+    for (long long i = start_i; i <= end_i; i++) {
       string num = to_string(i);
       int len = num.length();
+
       for (int j = 1; j <= len/2; j++) {
-        // cout << "B" << '\t' << j << endl;
         if (len % j != 0) continue;
+
         vector<string> parts;
         for (int k = j; k <= len; k+=j) {
           string s = num.substr(k-j,j);
           parts.push_back(s);
-          // cout << s << ' ';
         }
-        // cout << endl;
+
         bool b = true;
         for (string s : parts) {
           if (parts[0] != s) b = false;
         }
+
         if (b) {
           n += i;
-          // cout << "EUREKA: " << i << endl;
           break;
         }
       }
@@ -83,8 +86,8 @@ long part2() {
 }
 
 int main() {
-  long p1 = part1();
-  long p2 = part2();
+  long long p1 = part1();
+  long long p2 = part2();
   cout << "part 1: " << p1 << endl;
   cout << "part 2: " << p2 << endl;
   return 0;
